refactored SingleCohortStretchTables
